refactor(test): Reduce the code of TestThreadingModule by splitting up the code into smaller files.
