+++ app/src/androidTest/kotlin/ca/rmen/android/poetassistant/di/TestThreadingModule.kt
@@ -22,10 +22,9 @@ package ca.rmen.android.poetassistant.di
 import android.os.Handler
 import android.os.Looper
 import androidx.arch.core.executor.ArchTaskExecutor
-import androidx.arch.core.executor.DefaultTaskExecutor
 import androidx.test.espresso.IdlingRegistry
-import androidx.test.espresso.IdlingResource
-import androidx.test.espresso.idling.CountingIdlingResource
+import ca.rmen.android.poetassistant.testsupport.CountingIdlingResourceArchTaskExecutor
+import ca.rmen.android.poetassistant.testsupport.CountingIdlingResourceDispatcher
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.components.SingletonComponent
@@ -36,7 +35,6 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.android.asCoroutineDispatcher
 import kotlinx.coroutines.test.setMain
 import javax.inject.Singleton
-import kotlin.coroutines.CoroutineContext
 
 @OptIn(ExperimentalCoroutinesApi::class)
 @TestInstallIn(
@@ -46,66 +44,27 @@ import kotlin.coroutines.CoroutineContext
 @Module
 class TestThreadingModule {
 
-    private val testDispatcher = TrackingTestDispatcher("IO")
-    private val testMainDispatcher = TrackingTestDispatcher("Main",
-        Handler(Looper.getMainLooper()).asCoroutineDispatcher()
+    private val testIODispatcher = CountingIdlingResourceDispatcher(
+        "IO",
+        Dispatchers.IO,
     )
-    private val archIdlingResource = CountingIdlingResource("ArchIdlingResource", true)
+    private val testMainDispatcher = CountingIdlingResourceDispatcher(
+        "Main",
+        Handler(Looper.getMainLooper()).asCoroutineDispatcher(),
+    )
+    private val archExecutor = CountingIdlingResourceArchTaskExecutor()
 
     init {
-        IdlingRegistry.getInstance().register(testDispatcher.getIdlingResource())
-        IdlingRegistry.getInstance().register(testMainDispatcher.getIdlingResource())
-        IdlingRegistry.getInstance().register(archIdlingResource)
         Dispatchers.setMain(testMainDispatcher)
-        ArchTaskExecutor.getInstance().setDelegate(object : DefaultTaskExecutor() {
-            override fun executeOnDiskIO(runnable: Runnable) {
-                archIdlingResource.increment()
-                try {
-                    super.executeOnDiskIO(runnable)
-                } finally {
-                    archIdlingResource.decrementDelayed(200)
-                }
-            }
-
-            override fun postToMainThread(runnable: Runnable) {
-                archIdlingResource.increment()
-                try {
-                    super.postToMainThread(runnable)
-                } finally {
-                    archIdlingResource.decrementDelayed(200)
-                }
-            }
-        })
+        IdlingRegistry.getInstance().register(testIODispatcher.getIdlingResource())
+        IdlingRegistry.getInstance().register(testMainDispatcher.getIdlingResource())
+        IdlingRegistry.getInstance().register(archExecutor.getIdlingResource())
+        ArchTaskExecutor.getInstance().setDelegate(archExecutor)
     }
-
-    class TrackingTestDispatcher(
-        label: String,
-        private val delegate: CoroutineDispatcher = Dispatchers.IO
-    ) : CoroutineDispatcher() {
-
-        private val idlingResource = CountingIdlingResource("TrackingTestDispatcher-$label", true)
-
-        override fun dispatch(context: CoroutineContext, block: Runnable) {
-            idlingResource.increment()
-            delegate.dispatch(context) {
-                try {
-                    block.run()
-                } finally {
-                    idlingResource.decrementDelayed(200)
-                }
-            }
-        }
-
-        fun getIdlingResource(): IdlingResource = idlingResource
-    }
-
 
     @OptIn(ExperimentalCoroutinesApi::class)
     @Provides
     @Singleton
     @IODispatcher
-    fun providesIODispatcher(): CoroutineDispatcher = testDispatcher
-}
-private fun CountingIdlingResource.decrementDelayed(delay: Int) {
-    Handler(Looper.getMainLooper()).postDelayed({ decrement() }, 200)
+    fun providesIODispatcher(): CoroutineDispatcher = testIODispatcher
 }
