+++ app/src/androidTest/kotlin/ca/rmen/android/poetassistant/testsupport/CountingIdlingResourceExt.kt
@@ -1,0 +1,7 @@
+import android.os.Handler
+import android.os.Looper
+import androidx.test.espresso.idling.CountingIdlingResource
+
+fun CountingIdlingResource.decrementDelayed(delay: Long) {
+    Handler(Looper.getMainLooper()).postDelayed({ decrement() }, delay)
+}
